+++ spec/dummy/app/models/user.rb
@@ -4,7 +4,8 @@ class User
 
   field :name, type: String
 
-  attaches_one :photo, uploader: PictureUploader
+  attaches_one :photo,           uploader: PictureUploader
+  attaches_one :photo_and_video, uploader: PictureUploader, allow_videos: true
   attaches_one :compound_name_file
   attaches_one :file, type: :file
 end