+++ spec/lib/media_magick/controller/helper_spec.rb
@@ -1,26 +1,25 @@
 require 'spec_helper'
 
 describe MediaMagick::Controller::Helpers do
+  let(:controller) { Class.new }
+
   before do
-    @controller = Class.new
-    @controller.extend MediaMagick::Controller::Helpers
+    controller.extend MediaMagick::Controller::Helpers
   end
 
   describe "getting doc by params" do
-    before do
-      @track = Track.new
-      @album = Album.create(tracks: [@track])
-    end
+    let(:track) { Track.new }
+    let!(:album) { Album.create(tracks: [track]) }
 
     context "document is embedded" do
       it "should get parent by params" do
         params = {
           :embedded_in_model => "album", 
-          :embedded_in_id    => "#{@album.id.to_s}", 
+          :embedded_in_id    => "#{album.id.to_s}", 
           :model             => "track", 
-          :model_id          => "#{@track.id.to_s}"
+          :model_id          => "#{track.id.to_s}"
         }
-        @controller.find_doc_by_params(params).should eq(@track)
+        controller.find_doc_by_params(params).should eq(track)
       end
     end
 
@@ -28,9 +27,35 @@ describe MediaMagick::Controller::Helpers do
       it "should get parent by params" do
         params = {
           :model             => "album", 
-          :model_id          => "#{@album.id.to_s}"
+          :model_id          => "#{album.id.to_s}"
         }
-        @controller.find_doc_by_params(params).should eq(@album)
+        controller.find_doc_by_params(params).should eq(album)
+      end
+    end
+  end
+
+  describe "creating video" do
+    context "relation is attaches many" do
+      let(:video_url) { "youtube.com/watch?v=FfUHkPf9D9k" }
+      let!(:album) { Album.create }
+
+      it "should create video" do
+        params = {relation: "photos_and_videos", video: video_url}
+
+        controller.create_video(album, params).should be_true
+        album.reload.photos_and_videos.first.video.should eq(video_url)
+      end
+    end
+
+    context "relation is attaches one" do
+      let(:video_url) { "youtube.com/watch?v=FfUHkPf9D9k" }
+      let!(:user) { User.create }
+
+      it "should create video" do
+        params = {relation: "photo_and_video", video: video_url}
+
+        controller.create_video(user, params).should be_true
+        user.reload.photo_and_video.video.should eq(video_url)
       end
     end
   end